scripts/model_conversion/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
coremltools>=7.0
pillow>=9.0.0
timm>=0.9.0
numpy>=1.24.0
